:white_check_mark: re arrange helper tests closes #322
