+++ tests/helpers/isNumber.js
@@ -1,0 +1,44 @@
+'use strict';
+
+var assert = require('assert'),
+    helpers = require('../../lib/helpers');
+
+describe('helpers - isNumber', function () {
+
+  //////////////////////////////
+  // isNumber
+  //////////////////////////////
+
+  it('isNumber - [10 - true]', function (done) {
+
+    var result = helpers.isNumber(10);
+
+    assert.equal(true, result);
+    done();
+  });
+
+  it('isNumber - [\'10\' - true]', function (done) {
+
+    var result = helpers.isNumber('10');
+
+    assert.equal(true, result);
+    done();
+  });
+
+  it('isNumber - [\'ten\' - false]', function (done) {
+
+    var result = helpers.isNumber('ten');
+
+    assert.equal(false, result);
+    done();
+  });
+
+  it('isNumber - [\'ff00ff\' - false]', function (done) {
+
+    var result = helpers.isNumber('ff00ff');
+
+    assert.equal(false, result);
+    done();
+  });
+
+});
