+++ tests/helpers/sortDetects.js
@@ -1,0 +1,66 @@
+'use strict';
+
+var assert = require('assert'),
+    helpers = require('../../lib/helpers');
+
+var detectTestA = {
+      line: 1,
+      column: 1
+    },
+
+    detectTestB = {
+      line: 2,
+      column: 2
+    },
+
+    detectTestC = {
+      line: 1,
+      column: 2
+    };
+
+describe('helpers - sortDetects', function () {
+
+  //////////////////////////////
+  // sortDetects
+  //////////////////////////////
+
+  it('sortDetects should return -1 - [a.line < b.line]', function (done) {
+
+    var result = helpers.sortDetects(detectTestA, detectTestB);
+
+    assert.equal(-1, result);
+    done();
+  });
+
+  it('sortDetects should return 1 - [a.line > b.line]', function (done) {
+
+    var result = helpers.sortDetects(detectTestB, detectTestA);
+
+    assert.equal(1, result);
+    done();
+  });
+
+  it('sortDetects should return 0 - [a === b]', function (done) {
+
+    var result = helpers.sortDetects(detectTestA, detectTestA);
+
+    assert.equal(0, result);
+    done();
+  });
+
+  it('sortDetects should return -1 - [a.column < b.column]', function (done) {
+
+    var result = helpers.sortDetects(detectTestA, detectTestC);
+
+    assert.equal(-1, result);
+    done();
+  });
+
+  it('sortDetects should return 1 - [a.column > b.column]', function (done) {
+
+    var result = helpers.sortDetects(detectTestC, detectTestA);
+
+    assert.equal(1, result);
+    done();
+  });
+});
